+++ processor/src/test/resources/expected/com/example/deprecated/Arez_DeprecatedMemoizeModel.java
@@ -40,7 +40,7 @@ public final class Arez_DeprecatedMemoizeModel extends DeprecatedMemoizeModel im
     this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
     this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "DeprecatedMemoizeModel", $$arezi$$_id(), $$arezi$$_name(), null, null ) : null;
     this.$$arezi$$_disposedObservable = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".isDisposed" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> this.$$arezi$$_state >= 0 : null, null );
-    this.$$arez$$_count = new MemoizeCache<>( $$arezi$$_context(), Arez.areNamesEnabled() ? $$arezi$$_name() + ".count" : null, args -> super.count((long) args[ 0 ], (float) args[ 1 ]), 2);
+    this.$$arez$$_count = new MemoizeCache<>( $$arezi$$_context(), Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".count" : null, args -> super.count((long) args[ 0 ], (float) args[ 1 ]), 2);
     this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
     if ( Arez.areNativeComponentsEnabled() ) {
       this.$$arezi$$_component.complete();